Added threading and better comments
