+++ polus-precompute-slide-plugin/plugin.json
@@ -1,9 +1,14 @@
 {
   "name": "Polus Precompute Slide Plugin",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "title": "Precompute Slide Viewer",
   "description": "Precomputes a single image plane for viewing in Polus Volume.",
-  "containerId": "labshare/polus-precompute-slide-plugin:0.1.0",
+  "author": "Nick Schaub (nick.schaub@nih.gov)",
+  "institution": "National Center for Advancing Translational Sciences, National Institutes of Health",
+  "repository": "https://github.com/labshare/polus-plugins",
+  "website": "https://ncats.nih.gov/preclinical/core/informatics",
+  "citation": "",
+  "containerId": "labshare/polus-precompute-slide-plugin:0.2.0",
   "inputs": [
 
     {
